requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
bleach==3.3.1
boto3==1.17.29
botocore==1.20.29
celery==5.1.2
certifi==2020.12.5
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-admin-honeypot==1.1.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.0.0
django-ckeditor-5==0.0.12
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2.2
django-discover-runner==1.0
django-environ==0.7.0
django-eventstream==4.2.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grip==3.0.0
django-ipware==3.0.2
django-jet-reboot==1.1.1
django-js-asset==1.2.2
django-markdownfield==0.10.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-notifications-rest==0.0.1
django-phonenumber-field==5.0.0
django-role-permissions==3.1.1
django-rosetta==0.9.3
django-storages==1.11.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.2
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-recaptcha==2.0.4
drf-writable-nested==0.6.2
drf-yasg==1.20.0
flower==1.0.0
google-api-python-client==1.4.1
gripcontrol==4.1.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.19.1
humanize==3.11.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
phonenumbers==8.12.20
Pillow==8.1.0
polib==1.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.15.6
psycopg2==2.9.1
psycopg2-binary==2.8.6
pubcontrol==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
service-identity==21.1.0
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
swapper==1.1.2.post1
tornado==6.1
twilio==6.55.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.3.0
zope.interface==5.4.0
